Added a notebook with code examples. Updated package structure. Added a couple of tests for pandas functionality.
