+++ src/nimble_tk/__init__.py
@@ -1,3 +1,7 @@
 
 from .common import *
 from .notebook import *
+from .analytics.pandas_utils import *
+from .analytics.string_utils import *
+from .concurrent.tasks import *
+from .linux import *
