+++ src/nimble_tk/linux/__init__.py
@@ -1,7 +1,8 @@
 import zlib
 import datetime
 
-import common
+# import common
+from .. import common
 
 from typing import Optional, Callable, Tuple, Any
 import datetime
