+++ test/__init__.py
@@ -1,0 +1,3 @@
+"""
+This package contains automated code tests.
+"""
